Move traits related to datastructures to correct directory
- adding compatibility headers for those traits files
