+++ hpx/util/detail/container_category.hpp
@@ -9,7 +9,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/iterator_support/is_range.hpp>
-#include <hpx/traits/is_tuple_like.hpp>
+#include <hpx/datastructures/traits/is_tuple_like.hpp>
 
 namespace hpx {
 namespace util {
