+++ hpx/util/detail/unwrap_impl.hpp
@@ -10,7 +10,7 @@
 #include <hpx/config.hpp>
 #include <hpx/traits/future_traits.hpp>
 #include <hpx/traits/is_future.hpp>
-#include <hpx/traits/is_tuple_like.hpp>
+#include <hpx/datastructures/traits/is_tuple_like.hpp>
 #include <hpx/functional/invoke.hpp>
 #include <hpx/functional/invoke_fused.hpp>
 #include <hpx/util/pack_traversal.hpp>
