+++ libs/datastructures/include/hpx/datastructures/traits/supports_streaming_with_any.hpp
