+++ libs/datastructures/include_compatibility/hpx/traits/is_tuple_like.hpp
@@ -1,0 +1,19 @@
+//  Copyright (c) 2019 Auriane Reverdell
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#include <hpx/config.hpp>
+#include <hpx/datastructures/config/defines.hpp>
+#include <hpx/datastructures/traits/is_tuple_like.hpp>
+
+#if defined(HPX_DATASTRUCTURES_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_MSVC)
+#pragma message("The header hpx/traits/is_tuple_like.hpp is deprecated, \
+    please include datastructures/traits/is_tuple_like.hpp instead.")
+#else
+#warning "The header hpx/traits/is_tuple_like.hpp is deprecated, \
+    please include datastructures/traits/is_tuple_like.hpp instead."
+#endif
+#endif
