+++ libs/datastructures/tests/unit/is_tuple_like.cpp
@@ -9,9 +9,9 @@
 #include <vector>
 
 #include <hpx/config.hpp>
+#include <hpx/datastructures/traits/is_tuple_like.hpp>
 #include <hpx/datastructures/tuple.hpp>
 #include <hpx/testing.hpp>
-#include <hpx/traits/is_tuple_like.hpp>
 
 void tuple_like_true()
 {
